Update Riemann script to grab latest version
Now uses Clojars to grab the maven-metadata.xml file for the Riemann
project, runs an XPath query to grab all the versions, and filters out
any snapshots or other versions not matching the 'x.y.z' pattern.

Also adds a little bit more useful info/debug output as well.

On branch master
Your branch is up-to-date with 'origin/master'.

Changes to be committed:
  (use "git reset HEAD <file>..." to unstage)

	modified:   riemann.py